Clean-up warnings

+++ vpr/SRC/base/device_grid.cpp
@@ -25,8 +25,8 @@ void DeviceGrid::count_instances() {
     instance_counts_.clear();
 
     //Count the number of blocks in the grid
-    for(int x = 0; x < width(); ++x) {
-        for(int y = 0; y < width(); ++y) {
+    for(size_t x = 0; x < width(); ++x) {
+        for(size_t y = 0; y < width(); ++y) {
             auto type = grid_[x][y].type;
 
             if (grid_[x][y].width_offset == 0 && grid_[x][y].height_offset == 0) {
